requirements.txt
Django==4.0
djangorestframework==3.13.1
django-filter==21.1
#jwt
djangorestframework-simplejwt
#nested router
drf-nested-routers
#prostgres
psycopg2_binary==2.9.2
